Fixes for AST classes. Add parse tree and AST inspectors.
