+++ Src/js/cql-to-elm-ui/compiler/compiler.ts
@@ -2,7 +2,7 @@
 import * as ucum from "@lhncbc/ucum-lhc";
 import * as cqlToElmJs from "cql-to-elm-js";
 import * as cqlToElmWasmJs from "cql-to-elm-wasm-js";
-import { compilerOptions, TCompileCqlArgs, TOutput } from "@/shared";
+import { compilerOptions, TCqlToElmArgs, TOutput } from "@/shared";
 import { supportedModels } from "@/compiler/supported-models";
 import { fetchSync, readFile } from "@/compiler/utils";
 
@@ -55,7 +55,7 @@ export function createStatefulCompiler(sync: boolean) {
   );
 
   const compileCql = (
-    args: TCompileCqlArgs,
+    args: TCqlToElmArgs,
     onOutput: (output: TOutput) => void,
   ) => {
     const getModelXml = (
