+++ Src/js/cql-to-elm-ui/compiler/parse-tree.ts
@@ -1,0 +1,5 @@
+import { inspectCqlParseTree } from "cql-to-elm-js/kotlin/cql.mjs";
+
+export function cqlToParseTree(cql: string) {
+  return inspectCqlParseTree(cql);
+}
