+++ Src/js/cql-to-elm-ui/shared.tsx
@@ -4,6 +4,13 @@ export type TElmContentType = "json" | "xml";
 
 export type TLibrarySource = "local" | "remote";
 
+export type TMountedDir = {
+  handle: FileSystemDirectoryHandle;
+  files: {
+    handle: FileSystemFileHandle;
+  }[];
+};
+
 export const compilerOptions = [
   {
     value: "EnableDateRangeOptimization",
@@ -75,7 +82,7 @@ export const compilerOptions = [
 
 export const signatureLevels = ["None", "Differing", "Overloads", "All"];
 
-export type TCompileCqlArgs = {
+export type TCqlToElmArgs = {
   cql: string;
   useWasm: boolean;
   compilerOptions: string[];
@@ -83,12 +90,7 @@ export type TCompileCqlArgs = {
   outputContentType: TElmContentType;
   librarySource: TLibrarySource;
   baseUrl: string;
-  mountedDir: {
-    handle: FileSystemDirectoryHandle;
-    files: {
-      handle: FileSystemFileHandle;
-    }[];
-  } | null;
+  mountedDir: TMountedDir | null;
   useWorker: boolean;
 };
 
@@ -102,3 +104,58 @@ export type TOutput =
       contentType: TElmContentType;
       elm: string;
     };
+
+export const initialState = {
+  common: {
+    selectedTab: "cql-to-elm",
+    cql: `library Test version '0.1.0'
+
+using FHIR version '4.0.1'
+
+include FHIRHelpers version '4.0.1'
+
+valueset "Encounter Inpatient": 'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.666.5.307'
+
+parameter "Measurement Period" Interval<DateTime>
+
+context Patient
+
+define "Inpatient Encounter":
+  [Encounter: "Encounter Inpatient"] EncounterInpatient
+    where EncounterInpatient.status = 'finished'
+      and EncounterInpatient.period ends during day of "Measurement Period"
+`,
+    mountedDir: null as TMountedDir | null,
+    log: [] as string[],
+  },
+
+  tabs: {
+    "cql-to-parse-tree": {},
+    "cql-to-ast": {},
+    "cql-to-elm": {
+      cqlToElmArgs: {
+        useWasm: false,
+        compilerOptions: [
+          "EnableLocators",
+          "DisableListDemotion",
+          "DisableListPromotion",
+        ],
+        signatureLevel: "Overloads",
+        outputContentType: "json",
+        librarySource: "remote",
+        baseUrl:
+          "https://raw.githubusercontent.com/cqframework/cqf-exercises/refs/heads/master/input/cql/",
+        useWorker: true,
+      } as Omit<TCqlToElmArgs, "cql" | "mountedDir">,
+      elm: {
+        contentType: "json" as TElmContentType,
+        content: "",
+      },
+      isBusy: true,
+    },
+  },
+};
+
+export type TState = typeof initialState;
+
+export type TSetState = React.Dispatch<React.SetStateAction<TState>>;
